+++ src/Backend/SgateShopgatePlugin/Bootstrap.php
@@ -923,6 +923,17 @@ class Shopware_Plugins_Backend_SgateShopgatePlugin_Bootstrap extends Shopware_Co
                     ->setPosition($position++)
             );
 
+        $formElementCheckbox                 = new FormElementCheckbox();
+        $formElementOptionsContainerCheckbox = new FormElementOptionsContainerCheckbox();
+        $formElements[]                      = $formElementCheckbox
+            ->setKey('SGATE_EXPORT_PRODUCT_INACTIVE')
+            ->setOptions(
+                $formElementOptionsContainerCheckbox
+                    ->setLabel('Inaktive Produkte exportieren')
+                    ->setValue(false)
+                    ->setPosition($position++)
+            );
+
         $formElementSelect                 = new FormElementSelect();
         $formElementOptionsContainerSelect = new FormElementOptionsContainerSelect();
         $formElements[]                    = $formElementSelect